change all location var type from int to double and fix ball stuck bouncing  bug

+++ src/main/java/com/abstracts/GameObject.java
@@ -3,10 +3,10 @@ package com.abstracts;
 import javafx.scene.canvas.GraphicsContext;
 
 public abstract class GameObject {
-    private int x;
-    private int y;
-    private int width;
-    private int height;
+    private double x;
+    private double y;
+    private double width;
+    private double height;
 
     public GameObject() {
         this.x = 0;
@@ -15,7 +15,7 @@ public abstract class GameObject {
         this.height = 0;
     }
 
-    public GameObject(int x, int y, int width, int height) {
+    public GameObject(double x, double y, double width, double height) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -28,35 +28,35 @@ public abstract class GameObject {
     public void render(GraphicsContext gc) {
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 }
